+++ Tests/SimplePodcastManagerUITests/SyncStorageTestSupport.swift
@@ -4,9 +4,12 @@ import Foundation
 struct TestUISyncStorageInspector: SyncStorageInspecting {
     var availableBytes: Int64 = .max
     var fileSizeBytes: Int64 = 1
+    var sizesByPath: [String: Int64] = [:]
 
     func availableCapacity(on device: DeviceInfo) throws -> Int64 { availableBytes }
-    func fileSize(at url: URL) throws -> Int64 { fileSizeBytes }
+    func fileSize(at url: URL) throws -> Int64 {
+        sizesByPath[url.standardizedFileURL.path] ?? fileSizeBytes
+    }
 }
 
 func makeTestPlanner(